chore: move files to correct package and rename action
