+++ lib/features/immersion/search/immersion_filtres_middleware.dart
@@ -1,7 +1,7 @@
 import 'package:pass_emploi_app/features/immersion/list/immersion_list_actions.dart';
+import 'package:pass_emploi_app/features/immersion/search/immersion_search_parameters_actions.dart';
 import 'package:pass_emploi_app/features/immersion/search/immersion_search_parameters_state.dart';
 import 'package:pass_emploi_app/features/login/login_state.dart';
-import 'package:pass_emploi_app/redux/actions/immersion_actions.dart';
 import 'package:pass_emploi_app/redux/app_state.dart';
 import 'package:pass_emploi_app/repositories/immersion_repository.dart';
 import 'package:redux/redux.dart';
@@ -18,7 +18,7 @@ class ImmersionFiltresMiddleware extends MiddlewareClass<AppState> {
     final parametersState = store.state.immersionSearchParametersState;
     if (loginState is LoginSuccessState) {
       final userId = loginState.user.id;
-      if (action is ImmersionSearchUpdateFiltresAction &&
+      if (action is ImmersionSearchUpdateFiltresRequestAction &&
           parametersState is ImmersionSearchParametersInitializedState) {
         _resetSearchWithUpdatedFiltres(
           store: store,
