+++ lib/features/immersion/search/immersion_search_parameters_actions.dart
@@ -1,3 +1,5 @@
+import 'package:pass_emploi_app/models/immersion.dart';
+import 'package:pass_emploi_app/models/immersion_filtres_parameters.dart';
 import 'package:pass_emploi_app/repositories/immersion_repository.dart';
 
 class ImmersionSearchParametersRequestAction {
@@ -5,3 +7,17 @@ class ImmersionSearchParametersRequestAction {
 
   ImmersionSearchParametersRequestAction({required this.request});
 }
+
+class ImmersionSearchUpdateFiltresRequestAction {
+  final ImmersionSearchParametersFiltres updatedFiltres;
+
+  ImmersionSearchUpdateFiltresRequestAction(this.updatedFiltres);
+}
+
+class ImmersionSearchWithUpdateFiltresSuccessAction {
+  final List<Immersion> immersions;
+
+  ImmersionSearchWithUpdateFiltresSuccessAction({required this.immersions});
+}
+
+class ImmersionSearchWithUpdateFiltresFailureAction {}
