+++ lib/features/immersion/search/immersion_search_parameters_reducer.dart
@@ -2,7 +2,6 @@ import 'package:pass_emploi_app/features/immersion/list/immersion_list_actions.d
 import 'package:pass_emploi_app/features/immersion/search/immersion_search_parameters_actions.dart';
 import 'package:pass_emploi_app/features/immersion/search/immersion_search_parameters_state.dart';
 import 'package:pass_emploi_app/models/immersion_filtres_parameters.dart';
-import 'package:pass_emploi_app/redux/actions/immersion_actions.dart';
 
 ImmersionSearchParametersState immersionSearchParametersState(
   ImmersionSearchParametersState current,
@@ -16,7 +15,7 @@ ImmersionSearchParametersState immersionSearchParametersState(
       ville: '',
       title: action.request.title,
     );
-  } else if (action is ImmersionSearchUpdateFiltresAction) {
+  } else if (action is ImmersionSearchUpdateFiltresRequestAction) {
     if (current is ImmersionSearchParametersInitializedState) {
       return ImmersionSearchParametersInitializedState(
         location: current.location,
