+++ lib/presentation/immersion/immersion_filtres_view_model.dart
@@ -1,9 +1,9 @@
 import 'package:equatable/equatable.dart';
 import 'package:pass_emploi_app/features/immersion/list/immersion_list_state.dart';
+import 'package:pass_emploi_app/features/immersion/search/immersion_search_parameters_actions.dart';
 import 'package:pass_emploi_app/features/immersion/search/immersion_search_parameters_state.dart';
 import 'package:pass_emploi_app/models/immersion_filtres_parameters.dart';
 import 'package:pass_emploi_app/presentation/display_state.dart';
-import 'package:pass_emploi_app/redux/actions/immersion_actions.dart';
 import 'package:pass_emploi_app/redux/app_state.dart';
 import 'package:pass_emploi_app/ui/strings.dart';
 import 'package:redux/redux.dart';
@@ -66,5 +66,6 @@ int _distance(ImmersionSearchParametersState parametersState) {
 }
 
 void _dispatchUpdateFiltresAction(Store<AppState> store, int? updatedDistanceValue) {
-  store.dispatch(ImmersionSearchUpdateFiltresAction(ImmersionSearchParametersFiltres.distance(updatedDistanceValue)));
+  store.dispatch(
+      ImmersionSearchUpdateFiltresRequestAction(ImmersionSearchParametersFiltres.distance(updatedDistanceValue)));
 }
